move sortings util into sub-folder
