+++ src/utils/scoreColor.ts
@@ -2,14 +2,15 @@ import { Chalk } from 'chalk';
 import { getRenderColor, RenderColor } from '../cli/renderer/renderer';
 import { RatingAverage } from '../types/rating';
 
+const renderStates = {
+  [RatingAverage.Above]: getRenderColor(RenderColor.Success),
+  [RatingAverage.Neutral]: getRenderColor(RenderColor.Neutral),
+  [RatingAverage.Below]: getRenderColor(RenderColor.Error),
+};
+
 export const scoreColor = (score: number, average: number): Chalk => {
   let diff;
   const diffThreshold = 0.5;
-  const renderStates = {
-    [RatingAverage.Above]: getRenderColor(RenderColor.Success),
-    [RatingAverage.Neutral]: getRenderColor(RenderColor.Neutral),
-    [RatingAverage.Below]: getRenderColor(RenderColor.Error),
-  };
 
   if (score < average) {
     diff = average - score;
